requirements.txt
fastapi
pydantic
uvicorn
python-dotenv
sqlalchemy
python-jose
passlib[bcrypt]
python-multipart